requirements.txt
kafka-python==2.0.2
nltk==3.6.5
pymongo==3.12.1
requests==2.26.0
beautifulsoup4==4.10.0
pyspark==3.2.0
Flask==2.0.1
Werkzeug==2.0.1
elasticsearch==7.15.1
python-dotenv
feedparser